Fix '1000:Permission Denied' error [VPNLINUX-313] [VPNLINUX-314]
It seems that 'permission denied' is received when we  attempt to get settings from a 'protected' connection (that was possibly created by a 3rd party, ie: Docker or VMBox). The solution is to try and catch the exception and simply skip the connection since, in general, all user connections are accessible to the user.