Display proper message when unable to establish VPN tunnel [VPNLINUX-386]
Due to incorrect logic in how current vpn session where compared against max allowed vpn sessions, it created false positives when both values were equal, giving a false impression that a user has exceeded the ammount of available vpn sessions, while in fact it has not been exceeded.

This message is often tied with the fact when a machine is unable to establish a VPN tunnel, thus leading to possible confusions of why a user can not start a vpn connection.